+++ lib/matrix/subtract.go
@@ -7,9 +7,9 @@ func (m1 *Matrix) subtract(m2 *Matrix) *Matrix {
 
 	data1 := m1.Data
 	data2 := m2.Data
-	result := make([][]float32, len(data1))
+	result := make([][]float64, len(data1))
 	for i, row := range data1 {
-		result[i] = make([]float32, len(row))
+		result[i] = make([]float64, len(row))
 		for j := range row {
 			result[i][j] = data1[i][j] - data2[i][j]
 		}
@@ -18,10 +18,10 @@ func (m1 *Matrix) subtract(m2 *Matrix) *Matrix {
 
 }
 
-func (m *Matrix) singleSub(a float32) *Matrix {
-	result := make([][]float32, m.Rows)
+func (m *Matrix) singleSub(a float64) *Matrix {
+	result := make([][]float64, m.Rows)
 	for i := 0; i < m.Rows; i++ {
-		result[i] = make([]float32, m.Cols)
+		result[i] = make([]float64, m.Cols)
 		for j := 0; j < m.Cols; j++ {
 			result[i][j] = m.Data[i][j] - a
 		}
@@ -30,9 +30,9 @@ func (m *Matrix) singleSub(a float32) *Matrix {
 }
 
 func (m1 *Matrix) sub_with_lead_dim(m2 *Matrix) *Matrix {
-	result := make([][]float32, m1.Rows)
+	result := make([][]float64, m1.Rows)
 	for i, row := range m1.Data {
-		result[i] = make([]float32, m1.Cols)
+		result[i] = make([]float64, m1.Cols)
 		for j := range row {
 			result[i][j] = row[j] - m2.Data[0][j]
 		}
